requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
vaderSentiment
gtts
streamlit
newspaper3k
nltk
